// packages/ts/src/codegen/qorechain/svm/v1/query.ts
export interface QuerySlotResponse {
  slot: string;
}
// ...
export const QuerySlotResponse: MessageFns<QuerySlotResponse> = {
  encode(message: QuerySlotResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.slot !== "0") {
      writer.uint32(8).uint64(message.slot);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): QuerySlotResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseQuerySlotResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.slot = reader.uint64().toString();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): QuerySlotResponse {
    return { slot: isSet(object.slot) ? globalThis.String(object.slot) : "0" };
  },

  toJSON(message: QuerySlotResponse): unknown {
    const obj: any = {};
    if (message.slot !== "0") {
      obj.slot = message.slot;
    }
    return obj;
  },

  create(base?: DeepPartial<QuerySlotResponse>): QuerySlotResponse {
    return QuerySlotResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<QuerySlotResponse>): QuerySlotResponse {
    const message = createBaseQuerySlotResponse();
    message.slot = object.slot ?? "0";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}