// packages/ts/src/codegen/qorechain/rdk/v1/query.ts
export interface RollupView {
  rollupId: string;
  creator: string;
  profile: string;
  settlementMode: string;
  daBackend: string;
  blockTimeMs: string;
  maxTxPerBlock: string;
  vmType: string;
  status: string;
  stakeAmount: string;
  layerId: string;
  createdHeight: string;
}

/** BatchView is the query-facing view of a settlement batch. */
export interface BatchView {
  rollupId: string;
  batchIndex: string;
  stateRoot: string;
  prevStateRoot: string;
  txCount: string;
  dataHash: string;
  proofType: string;
  status: string;
  submittedAt: string;
  finalizedAt: string;
  withdrawalsRoot: string;
}

export interface QueryParamsRequest {
}

export interface QueryParamsResponse {
  params?: ParamsView | undefined;
}

export interface QueryRollupRequest {
  rollupId: string;
}

export interface QueryRollupResponse {
  rollup?: RollupView | undefined;
}

export interface QueryRollupsRequest {
}

export interface QueryRollupsResponse {
  rollups: RollupView[];
}

export interface QueryBatchRequest {
  rollupId: string;
  batchIndex: string;
}

export interface QueryBatchResponse {
  batch?: BatchView | undefined;
}

export interface QueryLatestBatchRequest {
  rollupId: string;
}

export interface QueryLatestBatchResponse {
  batch?: BatchView | undefined;
}
// ...
export const QueryRollupResponse: MessageFns<QueryRollupResponse> = {
  encode(message: QueryRollupResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.rollup !== undefined) {
      RollupView.encode(message.rollup, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): QueryRollupResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseQueryRollupResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.rollup = RollupView.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): QueryRollupResponse {
    return { rollup: isSet(object.rollup) ? RollupView.fromJSON(object.rollup) : undefined };
  },

  toJSON(message: QueryRollupResponse): unknown {
    const obj: any = {};
    if (message.rollup !== undefined) {
      obj.rollup = RollupView.toJSON(message.rollup);
    }
    return obj;
  },

  create(base?: DeepPartial<QueryRollupResponse>): QueryRollupResponse {
    return QueryRollupResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<QueryRollupResponse>): QueryRollupResponse {
    const message = createBaseQueryRollupResponse();
    message.rollup = (object.rollup !== undefined && object.rollup !== null)
      ? RollupView.fromPartial(object.rollup)
      : undefined;
    return message;
  },
};
// ...
export const QueryLatestBatchResponse: MessageFns<QueryLatestBatchResponse> = {
  encode(message: QueryLatestBatchResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.batch !== undefined) {
      BatchView.encode(message.batch, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): QueryLatestBatchResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseQueryLatestBatchResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.batch = BatchView.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): QueryLatestBatchResponse {
    return { batch: isSet(object.batch) ? BatchView.fromJSON(object.batch) : undefined };
  },

  toJSON(message: QueryLatestBatchResponse): unknown {
    const obj: any = {};
    if (message.batch !== undefined) {
      obj.batch = BatchView.toJSON(message.batch);
    }
    return obj;
  },

  create(base?: DeepPartial<QueryLatestBatchResponse>): QueryLatestBatchResponse {
    return QueryLatestBatchResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<QueryLatestBatchResponse>): QueryLatestBatchResponse {
    const message = createBaseQueryLatestBatchResponse();
    message.batch = (object.batch !== undefined && object.batch !== null)
      ? BatchView.fromPartial(object.batch)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}